nacl
Module secret_box

This module provides secret box pack and open functionality. It also provide ability to use cipner format with-nonce.

Modules

format_wn

Constants

JWK_ALG_NAME

NaCl secret box algorithm name for JWK’s (JSON Web Key)

KEY_LENGTH

Key length for NaCl’s boxes

NONCE_LENGTH

Nonce length for NaCl’s boxes

POLY_LENGTH

Length of Poly hash, used in NaCl’s boxes

Functions

open

This function opens xsalsa20+poly1305 formatted cipher, returning a message, trimmed of prefixed zeros, unlike crypto_secretbox_open from crypto_secretbox/xsalsa20poly1305/ref/box.c. Note also that cipher doesn’t have leading zeros, unlike C version.

pack

This function packs given message into xsalsa20+poly1305 secret-box bytes layout (trimmed of prefixed zeros, unlike crypto_secretbox from crypto_secretbox/xsalsa20poly1305/ref/box.c).