nacl
Module sign

Structs

Keypair

Constants

JWK_ALG_NAME

Signing algorithm name for JWK’s (JSON Web Key)

PUBLIC_KEY_LENGTH

Public key length for signing

SECRET_KEY_LENGTH

Secret key length for signing

SEED_LENGTH

Seed length, from which keypair is created

Functions

extract_pkey

Extracts public key from a given secret key

generate_keypair

Analog of crypto_sign_keypair in crypto_sign/ed25519/ref/keypair.c

open

Analog of crypto_sign_open in crypto_sign/ed25519/ref/open.c

sign

Analog of crypto_sign in crypto_sign/ed25519/ref/sign.c

signature

Signs message m with key sk, returning only signature bytes.

verify

Verifies that signature sig was created on message m with key with given public part pk.